feat: tabela de classificacao

+++ fifa_ufms/app/src/main/java/com/example/fifa_ufms/entities/Time.java
@@ -3,7 +3,6 @@ package com.example.fifa_ufms.entities;
 import androidx.room.ColumnInfo;
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
-import androidx.room.ForeignKey;
 
 @Entity(tableName = "Times")
 public class Time {
@@ -19,7 +18,7 @@ public class Time {
         this.corUniforme = corUniforme;
     }
 
-    public int getIdTime() {
+    public long getIdTime() {
         return idTime;
     }
